Add ConstantPool method to retrieve object by index
ConstantPool did not provide a method for to retrieve
and object from its pool of Constants via an index. A public method
that returns an object from the pool given an index value was added.

Most of the structures and fields that exist in a .class file refer to a
structure in the constant pool. For example:

    In the Class File structure, the member this_class holds an unsigned integer
    value that is a valid index into the constant pool (Java Docs, 4.1.0).
    The structure at that index is a CONSTANT_Class_info structure that contains
    the member name_index (Java Docs, 4.4.1). The value is an index into the
    constant pool. The structure at that index contains a UTF-8 string that
    represents the class or interface name (See Java Docs, 4.2.1). Because of this,
    the ConstantPool API needs to include a method for retrieving constant pool
    objects by index.

Reference:
Java Docs, https://docs.oracle.com/javase/specs/jvms/se19/html/jvms-4.html#jvms-4.4.1