Fix off by one error mapping indexes into constant pool
Objects in the constant pool that refer to another object in the pool
by index were being mapped incorrectly. The fix was to subtract 1 from
the index listed in the .class file, in order to reflect the correct
index in the cPool[] field of the ConstantPool class.

This fix was needed because the cPool[] array is indexed starting at 0,
however, in the jvm specification, a valid index to refer to in the
constant pool must be greater than 0 (Class File). This means that indexes need to be
transposed by subtracting 1 from the index indicated in the .class file.

This design decision was made so that cPool[0] is not unused.

References:
Class File, Java specification, https://docs.oracle.com/javase/specs/jvms/se19/html/jvms-4.html#jvms-4.4.10